src/orchestration/requirements.txt
# AG2 Orchestrator Dependencies
autogen-core>=0.2.0
autogen-agentchat>=0.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
websockets>=11.0
httpx>=0.24.0
